Move SchedulableMiningCoordinator to common package
